% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{.get_path}
\alias{.get_path}
\title{Generate the inst path}
\usage{
.get_path(filetype, n_tokens)
}
\arguments{
\item{filetype}{Character scalar; the type of file, like "lookup" or "vocab".}

\item{n_tokens}{Integer scalar; The number of tokens used for that file.}
}
\value{
Character scalar; the path to the file.
}
\description{
Generate the inst path
}
\keyword{internal}
