% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{.load_inst_rds}
\alias{.load_inst_rds}
\title{Load an RDS from inst Dir}
\usage{
.load_inst_rds(filetype, n_tokens)
}
\arguments{
\item{filetype}{Character scalar; the type of file, like "lookup" or "vocab".}

\item{n_tokens}{Integer scalar; The number of tokens used for that file.}
}
\value{
The R object.
}
\description{
Load an RDS from inst Dir
}
\keyword{internal}
