% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{morphemepiece_lookup}
\alias{morphemepiece_lookup}
\title{Load a Morphemepiece Lookup}
\usage{
morphemepiece_lookup()
}
\value{
A named character vector.
}
\description{
A morphemepiece lookup is a named character vector. The names of the vector
are the words, and the values are the space-separated morpheme breakdowns of
those words.
}
\examples{
head(morphemepiece_lookup())
}
