% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{morphemepiece_vocab}
\alias{morphemepiece_vocab}
\title{Load a Morphemepiece Vocabulary}
\usage{
morphemepiece_vocab()
}
\value{
A morphemepiece_vocabulary.
}
\description{
A morphemepiece vocabulary is a named integer vector with class
"morphemepiece_vocabulary". The names of the vector are the morphemes, and
the values are the integer identifiers of those tokens. The vocabulary is
0-indexed for compatibility with Python implementations.
}
\examples{
head(morphemepiece_vocab())
}
